+++ ArtvisDB/dbo/Tables/ModuleIssue.sql
@@ -1,14 +1,14 @@
 ﻿CREATE TABLE [dbo].[ModuleIssue] (
-    [moduleIssueID]     INT        IDENTITY (1, 1) NOT NULL,
-    [campaignID]        INT        NOT NULL,
-    [moduleID]          SMALLINT   NOT NULL,
-    [modulePricelistID] SMALLINT   NOT NULL,
-    [issueDate]         DATETIME   NOT NULL,
-    [rollerID]          INT        NOT NULL,
-    [ratio]             FLOAT (53) CONSTRAINT [DF_ModuleIssue_ratio] DEFAULT (1) NOT NULL,
-    [positionId]        SMALLINT   NOT NULL,
-    [isConfirmed]       BIT        CONSTRAINT [DF_ModuleIssue_isConfirmed] DEFAULT (0) NOT NULL,
-    [tariffPrice]       MONEY      CONSTRAINT [DF_ModuleIssue_tariffPrice] DEFAULT (0) NOT NULL,
+    [moduleIssueID]     INT              IDENTITY (1, 1) NOT NULL,
+    [campaignID]        INT              NOT NULL,
+    [moduleID]          SMALLINT         NOT NULL,
+    [modulePricelistID] SMALLINT         NOT NULL,
+    [issueDate]         DATETIME         NOT NULL,
+    [rollerID]          INT              NOT NULL,
+    [ratio]             DECIMAL (18, 10) CONSTRAINT [DF_ModuleIssue_ratio] DEFAULT ((1)) NOT NULL,
+    [positionId]        SMALLINT         NOT NULL,
+    [isConfirmed]       BIT              CONSTRAINT [DF_ModuleIssue_isConfirmed] DEFAULT (0) NOT NULL,
+    [tariffPrice]       DECIMAL (18, 2)  CONSTRAINT [DF_ModuleIssue_tariffPrice] DEFAULT ((0)) NOT NULL,
     CONSTRAINT [PK_ModuleIssue] PRIMARY KEY CLUSTERED ([moduleIssueID] ASC) WITH (FILLFACTOR = 90),
     CONSTRAINT [FK_ModuleIssue_Campaign] FOREIGN KEY ([campaignID]) REFERENCES [dbo].[Campaign] ([campaignID]) ON DELETE CASCADE,
     CONSTRAINT [FK_ModuleIssue_IssuePosition] FOREIGN KEY ([positionId]) REFERENCES [dbo].[iIssuePosition] ([positionId]),
@@ -16,6 +16,9 @@
     CONSTRAINT [FK_ModuleIssue_ModulePriceList] FOREIGN KEY ([modulePricelistID]) REFERENCES [dbo].[ModulePriceList] ([modulePriceListID]),
     CONSTRAINT [FK_ModuleIssue_Roller] FOREIGN KEY ([rollerID]) REFERENCES [dbo].[Roller] ([rollerID])
 );
+
+
+
 
 
 GO
